add helpers and partials

+++ controllers/api/comment-routes.js
@@ -2,19 +2,20 @@ const router = require('express').Router();
 const { Comment } = require('../../models');
 
 router.get('/', (req, res) => {
-    Comment.findAll()
-      .then(dbCommentData => res.json(dbCommentData))
-      .catch(err => {
-        console.log(err);
-        res.status(500).json(err);
-      });
-  });
-  
-  router.post('/', (req, res) => {
-    // expects => {comment_text: "This is the comment", user_id: 1, post_id: 2}
+  Comment.findAll()
+    .then(dbCommentData => res.json(dbCommentData))
+    .catch(err => {
+      console.log(err);
+      res.status(500).json(err);
+    });
+});
+
+router.post('/', (req, res) => {
+  // expects => {comment_text: "This is the comment", user_id: 1, post_id: 2}
+  if (req.session) {
     Comment.create({
       comment_text: req.body.comment_text,
-      user_id: req.body.user_id,
+      user_id: req.session.user_id,
       post_id: req.body.post_id
     })
       .then(dbCommentData => res.json(dbCommentData))
@@ -22,9 +23,11 @@ router.get('/', (req, res) => {
         console.log(err);
         res.status(400).json(err);
       });
-  });
-  
-  router.delete('/:id', (req, res) => {
+  }
+});
+
+router.delete('/:id', (req, res) => {
+  if (req.session) {
     Comment.destroy({
       where: {
         id: req.params.id
@@ -41,7 +44,7 @@ router.get('/', (req, res) => {
         console.log(err);
         res.status(500).json(err);
       });
-  });
-  
-  module.exports = router;
-  
+  }
+});
+
+module.exports = router;
